+++ source/Examples/Resources/SystemProtection/1-SystemProtection_EnableDriveC_Config.ps1
@@ -1,0 +1,37 @@
+<#PSScriptInfo
+.VERSION 1.0.0
+.GUID 4f0d0a70-30e3-4f16-86f3-76631587bdd0
+.AUTHOR DSC Community
+.COMPANYNAME DSC Community
+.COPYRIGHT Copyright the DSC Community contributors. All rights reserved.
+.TAGS DSCConfiguration
+.LICENSEURI https://github.com/dsccommunity/ComputerManagementDsc/blob/master/LICENSE
+.PROJECTURI https://github.com/dsccommunity/ComputerManagementDsc
+.ICONURI
+.EXTERNALMODULEDEPENDENCIES
+.REQUIREDSCRIPTS
+.EXTERNALSCRIPTDEPENDENCIES
+.RELEASENOTES First version.
+.PRIVATEDATA 2016-Datacenter,2016-Datacenter-Server-Core
+#>
+
+#Requires -module ComputerManagementDsc
+
+<#
+    .DESCRIPTION
+        Enables system protection for the C drive using the
+        default value of 10 percent disk usage.
+#>
+Configuration SystemProtection_EnableDriveC_Config
+{
+    Import-DSCResource -ModuleName ComputerManagementDsc
+
+    Node localhost
+    {
+        SystemProtection DriveC
+        {
+            Ensure      = 'Present'
+            DriveLetter = 'C'
+        }
+    }
+}